Update for dbt2 v0.46.0
The client was refactored to build support for all databases that
libraries were detected for.

Packaging was updated to reflact that, meaning only 1 client package
is now needed.

For some reason, DBT-2 does not build correctly with MySQL and unixODBC
libraries on RHEL8 based systems, so support is not built for them in
these RPMs at this time.